Fix 32-bit host: accept int64 epoch seconds; enlarge Toit stack (#3076)
## Summary
Two related fixes for intermittent failures in the 32-bit Debug
nightly CI (pkg gold tests).

### `file.update-time`: accept int64 seconds
The `update_times` primitive declared `atime_sec` / `mtime_sec` as
`int`, which in the ARGS macro only accepts Smi. Smi max on 32-bit
is 2^30 - 1 ≈ 1.07e9, but the current Unix epoch is ≈ 1.78e9, so
`Time.now.s-since-epoch` is a `LargeInteger` and the primitive
rejected it with `OUT_OF_RANGE`.

### 32-bit host page size matches 64-bit
A Toit stack's max length is bounded by `TOIT_PAGE_SIZE`, which was
4kB on every 32-bit build vs 32kB on 64-bit — an 8x smaller stack
cap. On constrained devices (FreeRTOS) we want to keep 4kB, but a
32-bit host build has plenty of memory. The smaller cap was tipping
`yaml.decode` over the edge when it ran deep inside the pkg solver,
showing up as the intermittent "Failed to decode <path> as YAML"
`STACK_OVERFLOW` in nightly 32-bit Debug.

Now `TOIT_PAGE_SIZE_LOG2` only falls back to the 32-bit value when
`TOIT_FREERTOS` is defined.

+++ src/primitive_file_non_win.cc
@@ -409,7 +409,9 @@ const long kToitNow = (1l << 30) - 1l;  // Special value for now.
 const long kToitNoTime = (1l << 30) - 2l;  // Special value for no time change.
 
 PRIMITIVE(update_times) {
-  ARGS(cstring, path, int, atime_sec, int, atime_nsec, int, mtime_sec, int, mtime_nsec);
+  // Seconds-since-epoch exceed Smi range on 32-bit hosts (Smi max is ~1.07e9,
+  // current epoch is ~1.78e9), so we take them as int64.
+  ARGS(cstring, path, int64, atime_sec, int, atime_nsec, int64, mtime_sec, int, mtime_nsec);
 #ifndef TOIT_FREERTOS
   if (atime_nsec == kToitNow) {
     if (atime_sec != 0) FAIL(INVALID_ARGUMENT);
